requirements.txt
google-generativeai
tqdm
